Remove redundant saved-outfits.js file.
